Add watch parameter and improve logging (#34)
* Simplify tailwind downloader

* Add watch mode parameter to TailwindCSS build

This commit adds a new `Watch` property to the `BuildTailwindCSS` task, allowing users to run TailwindCSS in watch mode. The `Watch` property is set in the MSBuild target file and passed as an argument to the CLI command. Additionally, the commit updates the initialization command in case the configuration file doesn't exist.